requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
asyncio==3.4.3
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.2.0
blinker==1.8.2
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==3.4.0
click==8.1.7
cloudevents==1.9.0
comm==0.2.3
contourpy==1.3.3
cryptography==41.0.3
cycler==0.12.1
db-dtypes==1.3.0
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
dnspython==2.7.0
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.115.8
fastjsonschema==2.21.2
firebase-admin==5.2.0
Flask==2.3.3
fonttools==4.59.0
fqdn==1.5.1
frozenlist==1.5.0
functions-framework==3.0.0
google-api-core==2.22.0
google-api-python-client==2.184.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.24.0
google-cloud-bigquery-storage==2.29.1
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-pubsub==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.14.0
grpcio==1.67.1
grpcio-status==1.67.1
gspread==6.2.1
gunicorn==20.1.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
intervaltree==3.1.0
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
jwt==1.3.1
kiwisolver==1.4.9
lark==1.3.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mistune==3.1.4
msal==1.33.0
msgpack==1.1.0
multidict==6.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.7
notebook_shim==0.2.4
numpy==1.23.2
oauthlib==3.3.1
openpyxl==3.1.2
overrides==7.7.0
packaging==24.1
pandas==1.5.0
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.2.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.28.3
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
PyJWT==2.6.0
pymongo==4.4.1
pyparsing==3.2.0
pytest==7.4.0
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2024.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
rsa==4.9
scipy==1.16.1
seaborn==0.13.2
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
stack-data==0.6.3
starlette==0.45.3
stripe==11.5.0
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
twilio==8.8.0
types-python-dateutil==2.9.0.20250822
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
watchdog==1.0.2
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.2
widgetsnbextension==4.0.14
yarl==1.18.3
